backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
langgraph>=0.0.40
langchain-core>=0.1.0
redis>=5.0.0
chromadb==0.4.18
sentence-transformers>=2.3.0
pdfplumber==0.10.3
unstructured==0.11.8
supabase==2.0.0
httpx>=0.24.0,<0.25.0
huggingface-hub>=0.19.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
pytest==7.4.3
pytest-asyncio==0.21.1